requirements.txt
tensorflow==2.13.0
opencv-python
numpy
matplotlib
pandas
Pillow
scikit-learn